Remove unused BlockStateProviderMixin
This was entirely unused, as far as I'm aware. Originally I included it because Nature's Spirit used it and I didn't know for what so I just threw it in there during that process and I forgot to remove it. Anywho, got scared  because LTONES having the same mixin caused a crash, so this will hopefully avoid an issue like that involving nature's spirit.